Ancora/requirements.txt
openpyxl
pdfplumber
python-docx
PyPDF2
